requirements.txt
numpy==1.24.3
streamlit==1.29.0
Pillow==10.1.0
pandas==2.0.3
matplotlib==3.7.2
ultralytics==8.0.196
torch==2.1.0
torchvision==0.16.0
opencv-python-headless==4.10.0.84
python-dateutil==2.8.2 